Initial commit.
The app currently supports numerous accounts and budgets as well as
transactions. Transactions may be applied to one account and one budget.
The balance of the account is kept up to date by the date of the
transactions. Changes to amounts or debit/credit update the balance and
each transaction balance effected by the change.